Clean up FFI import
