+++ python/src/walnuts/__init__.py
@@ -1,4 +1,4 @@
-from .ffi import logp_cfunc_type, walnuts_pyfunc
+from .pyfunc import logp_cfunc_type, walnuts_pyfunc
 
 try:
     from .stan import walnuts_stan
@@ -9,7 +9,7 @@ except ImportError as e:
             "Failed to load walnuts_stan. Is bridgestan installed and working?"
         ) from e
 
-from .summary import Summarizer, r_hat, ess, mcse
+from .summary import Summarizer, ess, mcse, r_hat
 
 __all__ = ["walnuts_stan", "walnuts_pyfunc", "logp_cfunc_type", "r_hat", "ess", "mcse", "Summarizer"]
 __version__ = "0.0.1"
